+++ apps/forge-web/src/apps/Explorer.js
@@ -4,11 +4,7 @@ import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-d
 import { IntlProvider, addLocaleData } from 'react-intl';
 
 import AsyncComponent from '../components/async';
-import Layout from '../layouts/explorer';
 import ActivityIndicator from '../components/activity_indicator';
-
-import PageHome from '../pages/explorer/home';
-import PageAnimation from '../pages/explorer/animation';
 
 import { localeData } from '../libs/locale';
 import { detectLocale } from '../libs/util';
@@ -18,6 +14,8 @@ addLocaleData(localeData);
 
 const { locale, messages } = detectLocale();
 
+const Layout = AsyncComponent(() => import('../layouts/explorer'));
+const PageHome = AsyncComponent(() => import('../pages/explorer/home'));
 const PageNetwork = AsyncComponent(() => import('../pages/explorer/network'));
 const PageBlockList = AsyncComponent(() => import('../pages/explorer/blocks'));
 const PageBlockDetail = AsyncComponent(() => import('../pages/explorer/block'));
@@ -61,7 +59,6 @@ const App = () => {
             <Route exact path="/accounts/:address" component={PageAccountDetail} />
             <Route exact path="/assets/:address" component={PageAssetDetail} />
             <Route exact path="/network" component={PageNetwork} />
-            <Route exact path="/switcher" component={PageAnimation} />
             <Redirect from="/*" to="/" />
           </Switch>
         </Layout>