+++ apps/forge-web/src/layouts/explorer/index.js
@@ -6,7 +6,8 @@ import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 
 import ThemeSwitcher from '../../components/theme_switcher';
-import SearchBox from '../../components/search_box';
+import NetworkSwitcher from '../../components/network_switcher';
+
 import withI18n from '../../components/withI18n';
 import withRoot from '../../components/withRoot';
 import withTracker from '../../components/withTracker';
@@ -24,8 +25,10 @@ function Layout({ children }) {
       <AppBar position="absolute" className="appBar">
         <Toolbar disableGutters={false} className="toolbar">
           <Logo />
-          <SearchBox className="search-box" />
-          <ThemeSwitcher className="switcher" />
+          <div className="switchers">
+            <ThemeSwitcher className="switcher" />
+            <NetworkSwitcher className="switcher" />
+          </div>
         </Toolbar>
       </AppBar>
       <main className="main">
@@ -82,14 +85,15 @@ const Container = styled.div`
     margin-top: ${props => props.theme.spacing.unit}px;
     width: 100%;
     max-width: ${props => props.theme.pageWidth}px;
-    position: relative;
+
+    .switchers {
+      display: flex;
+      justify-content: flex-end;
+      align-items: center;
+    }
 
     .switcher {
       cursor: pointer;
-      @media (min-width: ${props => props.theme.breakpoints.values.lg}px) {
-        position: absolute;
-        right: -48px;
-      }
     }
   }
 
@@ -99,12 +103,6 @@ const Container = styled.div`
     overflow: auto;
     box-sizing: border-box;
     position: relative;
-  }
-
-  @media (max-width: ${props => props.theme.breakpoints.values.sm}px) {
-    .search-box {
-      display: none;
-    }
   }
 `;